refactor(DismissalCrewSortPolicy): dto의 정렬 로직을 담당하므로 dto 패키지로 이동
