+++ src/test/java/model/AttendanceBookTest.java
@@ -132,4 +132,3 @@ class AttendanceBookTest {
         );
     }
 }
-
